🚚 Rename `index.jsx` file in `src/routes/` folder to `Router.jsx`
Since the folder is now named `routes` instead of `Router` like the main component, the file should be renamed accordingly for better clarity.
